+++ extrasafe/examples/no_files_allow_stdout.rs
