+++ extrasafe/examples/simple_network.rs
