+++ extrasafe/examples/time.rs
