+++ extrasafe/src/macros.rs
