+++ extrasafe/tests/bad_combination.rs
