+++ extrasafe/tests/open_readonly.rs
